requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
httpx==0.28.1
feedparser==6.0.11
trafilatura==2.0.0
readability-lxml==0.8.4.1
beautifulsoup4==4.13.3
pyodbc==5.2.0
azure-identity==1.20.0
openai==1.65.4
sqlglot==26.10.1
pytest==8.3.5
pytest-cov==6.0.0
ruff==0.9.10
